chore(text field): extract helper text to separate component
