+++ packages/react/src/components/Input/Input.props.ts
@@ -20,6 +20,10 @@ export interface BaseProps {
    */
   id: string
   /**
+   * The helper text content.
+   */
+  helperText?: string
+  /**
    * If filled, will appear above the field.
    */
   label?: string
