+++ packages/react/src/components/Input/InputHelperText/InputHelperText.props.ts
@@ -1,0 +1,3 @@
+import { InputProps } from '../Input.props'
+
+export type InputHelperTextProps = Pick<InputProps, 'feedback' | 'helperText' | 'disabled'>
