+++ packages/react/src/components/Input/InputHelperText/index.ts
@@ -1,0 +1,4 @@
+import InputHelperText from './InputHelperText'
+
+export { InputHelperText }
+export default InputHelperText
