+++ packages/react/src/components/TextField/TextField.props.ts
@@ -2,10 +2,6 @@ import { InputProps } from 'components/Input'
 
 type BaseProps = InputProps & {
   /**
-   * The helper text content.
-   */
-  helperText?: string
-  /**
    * Optional ID for testing
    */
   testID?: string
